Fix selection validation error
The isDayDisabledFn was being called with a non-floored (not beginning of day) Posix during validation of the selection. Because an implementor should assume that the Posix provided to this function represents the beginning of a day, this led to the possibility of an edgecase in which the isDayDisabledFn would return True (disabled) and the validation would fail despite the day being rendered as enabled on the picker.